Renamed package, optimized imports, moved names map to global scope, changed 1 test
